Added Errorhandling to FileSystem

+++ enginesrc/IO/FileSystem.cpp
@@ -2,24 +2,25 @@
 
 #include <boost/filesystem.hpp>
 
+#include "Core/Singleton.h"
+#include "SubManager/ErrorHandler.h"
+
 namespace le{
 	namespace io{
 		bool FileSystem::copyFile( const std::string& path, const std::string& outPath, const bool overwrite )
 		{
-			//Todo: Error handling
-			if( fileExist( path ) ){
-				if( fileExist( outPath ) ){
-					if( overwrite ){
-						boost::filesystem::copy_file( path, outPath, boost::filesystem::copy_option::overwrite_if_exists );
-						return true;
-					} else{
-						return false;
-					}
-				} else{
-					boost::filesystem::copy_file( path, outPath );
+			if( fileExist( path ) && fileExist( outPath ) ){
+				if( overwrite ){
+					boost::filesystem::copy_file( path, outPath, boost::filesystem::copy_option::overwrite_if_exists );
 					return true;
+				} else{
+					lerr_setErr( sm::LERR_NOPRIVILEGE );
+					return false;
 				}
+			} else if( fileExist( path ) && !fileExist( outPath ) ){
+				boost::filesystem::copy_file( path, outPath );
 			} else{
+				lerr_setErr( sm::LERR_FILENOTFOUND );
 				return false;
 			}
 		}
@@ -30,7 +31,7 @@ namespace le{
 				std::remove( path.c_str( ) );
 				return true;
 			}
-			//Todo: Error Handling
+			lerr_setErr( sm::LERR_FILENOTFOUND );
 			return false;
 		}
 
@@ -41,81 +42,119 @@ namespace le{
 
 		bool FileSystem::readFile( const std::string& path, std::string& outStr )
 		{
-			//Todo: Error Handling
-			boost::filesystem::path p{ path };
-			boost::filesystem::ifstream ifs{ p };
-			std::string temp;
-			while( getline( ifs, temp ) )
-				outStr += temp + '\n';
-			return true;
+			if( fileExist( path ) ){
+				//Todo: what if File is already opened?(dont know if thats a problem while reading)
+				boost::filesystem::path p{ path };
+				boost::filesystem::ifstream ifs{ p };
+				std::string temp;
+				while( getline( ifs, temp ) )
+					outStr += temp + '\n';
+				return true;
+			} else{
+				lerr_setErr( sm::LERR_FILENOTFOUND );
+				return false;
+			}
 		}
 
 		std::string FileSystem::readFile( const std::string& path )
 		{
-			//Todo: Error Handling
-			boost::filesystem::path p{ path };
-			boost::filesystem::ifstream ifs{ p };
-			std::string temp, a;
-			while( getline( ifs, a ) )
-				temp += a + '\n';;
-			return temp;
+			if( fileExist( path ) ){
+				boost::filesystem::path p{ path };
+				boost::filesystem::ifstream ifs{ p };
+				std::string temp, a;
+				while( getline( ifs, a ) )
+					temp += a + '\n';;
+				return temp;
+			} else{
+				lerr_setErr( sm::LERR_FILENOTFOUND );
+				return "";
+			}
 		}
 
 		bool FileSystem::readFileRegionF( const std::string & path, std::string & outStr, int begin )
 		{
-			outStr = readFile( path ).substr( begin );
-			//Todo: Error Handling
-			return true;
+			if( fileExist( path ) ){
+				outStr = readFile( path ).substr( begin );
+				return true;
+			} else{
+				lerr_setErr( sm::LERR_FILENOTFOUND );
+				return false;
+			}
 		}
 
 		std::string FileSystem::readFileRegionF( const std::string & path, int begin )
 		{
-
-			return readFile( path ).substr( begin );
+			if( fileExist( path ) ){
+				return readFile( path ).substr( begin );
+			} else{
+				lerr_setErr( sm::LERR_FILENOTFOUND );
+				return "";
+			}
 		}
 
 		bool FileSystem::readFileRegionFT( const std::string & path, std::string & outStr, int begin, int len )
 		{
-
-			outStr = readFile( path ).substr( begin, len );
-			//Todo: Error Handling
-			return true;
+			if( fileExist( path ) ){
+				outStr = readFile( path ).substr( begin, len );
+				return true;
+			} else{
+				lerr_setErr( sm::LERR_FILENOTFOUND );
+				return false;
+			}
 		}
 
 		std::string FileSystem::readFileRegionFT( const std::string & path, int begin, int len )
 		{
-			return readFile( path ).substr( begin, len );
+			if( fileExist( path ) ){
+				return readFile( path ).substr( begin, len );
+			} else{
+				lerr_setErr( sm::LERR_FILENOTFOUND );
+				return "";
+			}
 		}
 
 		bool FileSystem::writeFile( const std::string& path, const std::string& inStr )
 		{
-			//Todo: Error Handling
-			//Todo: Add Overwrite Handling
-			boost::filesystem::path p{ path };
-			std::string temp = readFile( path );
-			boost::filesystem::ofstream ofs{ p };
-			ofs << temp << inStr;
-			return true;
+			if( fileExist( path ) ){
+				//Todo: Add Overwrite Handling
+				//Todo: what if file is already opened?
+				boost::filesystem::path p{ path };
+				std::string temp = readFile( path );
+				boost::filesystem::ofstream ofs{ p };
+				ofs << temp << inStr;
+				return true;
+			} else{
+				lerr_setErr( sm::LERR_FILENOTFOUND );
+				return false;
+			}
 		}
 		bool FileSystem::overwriteFile( const std::string & path, const std::string & inStr )
 		{
-
-			//Todo: Error Handling
-			//Todo: Add Overwrite Handling
-			boost::filesystem::path p{ path };
-			boost::filesystem::ofstream ofs{ p };
-			ofs <<inStr;
-			return true;
+			if( fileExist( path ) ){
+				//Todo: Add Overwrite Handling
+				//Todo: what if file is already opened?
+				boost::filesystem::path p{ path };
+				boost::filesystem::ofstream ofs{ p };
+				ofs << inStr;
+				return true;
+			} else{
+				lerr_setErr( sm::LERR_FILENOTFOUND );
+				return false;
+			}
 		}
 		bool FileSystem::writeFileOffset( const std::string & path, const std::string & inStr, int offset )
 		{
-			//Todo: Error Handling
-			//Todo: Add Overwrite Handling
-			boost::filesystem::path p{ path };
-			std::string temp = readFileRegionFT( path ,0,offset) + inStr + readFileRegionF(path,offset);
-			boost::filesystem::ofstream ofs{ p };
-			ofs << temp;
-			return true;
+			if( fileExist( path ) ){
+				//Todo: Add Overwrite Handling
+				boost::filesystem::path p{ path };
+				std::string temp = readFileRegionFT( path, 0, offset ) + inStr + readFileRegionF( path, offset );
+				boost::filesystem::ofstream ofs{ p };
+				ofs << temp;
+				return true;
+			} else{
+				lerr_setErr( sm::LERR_FILENOTFOUND );
+				return false;
+			}
 		}
 	}
 }
